Made the Zig sprite API match Python, turned disabled-feature calls into compile errors, and fixed vsync and Python safety issues
Sprite gained setPosF, setSize, setScale, setSrcRect (pixels), pos(), scale(), and a tint field that drawSprite routes on. ResourceManager.createSprite(name) builds one by name. addSubTexture now takes pixel RectI relative to the parent frame; the UV form is addSubTextureUV. Fixed the Sprite.create doc comment.

Renderer text/shape methods and eng.audio (DisabledAudioEngine) are compile errors naming the option when that feature is compiled out. Removed projMat, uiMatrix, and gameScale; examples use projection() and logicalSize. vsyncEnabled = false now disables vsync.

C API: fixed the python-ffi build broken by the Sprite.create change, validated key/button/axis ints instead of @enumFromInt panics, and float sprite positions. Python: update returning None keeps running, wrappers are marked destroyed when run() shuts the engine down, and run() refuses to start twice.

+++ python/pixzig/_native.py
@@ -154,7 +154,7 @@ pz_set_default_font = _sig("pz_set_default_font", [PzEnginePtr, c_char_p], c_int
 
 # --- Sprites -----------------------------------------------------------
 pz_sprite_create = _sig("pz_sprite_create", [PzEnginePtr, c_char_p], PzSpritePtr)
-pz_sprite_set_pos = _sig("pz_sprite_set_pos", [PzSpritePtr, c_int32, c_int32], None)
+pz_sprite_set_pos = _sig("pz_sprite_set_pos", [PzSpritePtr, c_float, c_float], None)
 pz_sprite_get_rect = _sig(
     "pz_sprite_get_rect",
     [PzSpritePtr, ctypes.POINTER(c_float), ctypes.POINTER(c_float), ctypes.POINTER(c_float), ctypes.POINTER(c_float)],
